[refactor/#28] 파일 이동 및 주석처리
