+++ src/main/java/com/goormthon/backend/firstsori/domain/board/application/dto/response/BoardInfoResponse.java
@@ -1,4 +1,4 @@
-package com.goormthon.backend.firstsori.domain.board.application.dto;
+package com.goormthon.backend.firstsori.domain.board.application.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
